+++ DelvUI/changelog.md
@@ -1,8 +1,9 @@
 # 2.3.1.1
 - Fixed Party Frames preview mode.
+- Fixed Enemy Nameplates.
 
 # 2.3.1.0
-- Fixed serveral issues with the Party Frames and Party Cooldowns.
+- Fixed several issues with the Party Frames and Party Cooldowns.
 - Fixed Ready Check Status icon in the Party Frames.
 - Removed the Player Order Override setting from Party Frames:
   * This is now possible through the game's systems.